// @types/csvData.d.ts
export interface csvData {
  uniqueName: string;
  jsonDataDE?: string;
  jsonDataEN?: string;
  jsonDataES?: string;
  jsonDataFR?: string;
  jsonDataIT?: string;
  jsonDataJA?: string;
  jsonDataKO?: string;
  jsonDataPL?: string;
  jsonDataPT?: string;
  jsonDataRU?: string;
  jsonDataTH?: string;
  jsonDataTR?: string;
  jsonDataUK?: string;
  jsonDataZH?: string;
  jsonDataTC?: string;
}